// src/app/Models/AuthorDetailsModel.ts
export interface AuthorDetails {
    name: string;
    url: string;
    sort_key: string;
    description: string;
    description_pl: string;
    plural: string;
    genre_epoch_specific: boolean;
    adjective_feminine_singular: string;
    adjective_nonmasculine_plural: string;
    genitive: string;
    collective_noun: string;
}